requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
asgiref==3.11.1
attrs==25.4.0
discord.py==2.7.1
Django==6.0.3
django-cors-headers==4.9.0
django-crontab==0.7.1
djangorestframework==3.16.1
frozenlist==1.8.0
idna==3.11
multidict==6.7.1
pillow==12.1.1
propcache==0.4.1
python-decouple==3.8
sqlparse==0.5.5
typing_extensions==4.15.0
yarl==1.23.0
